+++ src/ti/ttask.c
@@ -1555,6 +1555,7 @@ int ti_ttask_run(ti_change_t * change, mp_unp_t * up)
     case TI_TASK_THING_REMOVE:      break;
     case TI_TASK_SET_DEFAULT_DEEP:  return ttask__set_default_deep(up);
     case TI_TASK_RESTORE_FINISHED:  return ttask__restore_finished(up);
+    case TI_TASK_TO_THING:          break;
     }
 
     log_critical("unknown thingsdb task: %"PRIu64, mp_task.via.u64);